Get rid of negated array creation by making two near-copies of trees.
There are now an increasing tree and a decreasing tree. Their code
is largely the same except for four locations. I made several attempts
to make an implementation without code duplicates and failed.
The current approach is to some extent ugly,
but it is faster by ~10% than the previous version.